Tooltip for hex code copy (#35)
* Tooltip for hex code copy.

Now shows glowing red icon next to the hexcode when the user first visits the page. Clicking on this will tell users that they can click on the hex code to copy it. Uses React Joyride.

* react-joyride documented in package.json

* Tooltip properly follows hexcode

Tooltip will now always display when a page comes up with a hexcode. Joyride is keyed based on levelNum so that it updates properly based on which page we're on (no longer requires a refresh).

* Tooltip text centered vertically, opens on hover, and appears only on Level 2

* Done

* Fixed known hexcode click before hover on tooltip bug.

Now, the site doesn't crash if users first click the hexcode to copy before hoverin\ g over the tooltip (an unlikely situation but good to get rid of possibility).

Implementation notes: properly uses React Joyride's callback for when tooltip state\ changes. See https://docs.react-joyride.com/callback

* Fixed merge conflicts

Did not realize Celebi's tooltips were not at the top of the return in Main, my bad.

Co-authored-by: Nhamakami7 <nicholas_hamakami7@outlook.com>