Level 5 Descendants (#56)
* Level tutorial for CSS descendant selectors

* HTML descendants in Level 5 now formatted properly.

Changed the structure of the htmlContent Array. The last field for each element in that array now tracks the number of children each element has. This allows us to differentiate between true parents (parents that have children) and what I call "non-true" parents (HTML elements that don't have any children).

There are special edge cases handled with special numbers (-1 and -2) to indicate a true child (child of an element that doesn't have children itself) and closing tags (tags that close a previously opened parent element). See code for more details on this.

* HTML descendants in Level 5 now formatted properly.

Changed the structure of the htmlContent Array. The last field for each element in that array now tracks the number of children each element has. This allows us to differentiate between true parents (parents that have children) and what I call "non-true" parents (HTML elements that don't have any children).

There are special edge cases handled with special numbers (-1 and -2) to indicate a true child (child of an element that doesn't have children itself) and closing tags (tags that close a previously opened parent element). See code for more details on this.

* Changed /> to /&gt

Co-authored-by: Nhamakami7 <nicholas_hamakami7@outlook.com>